Refactor stage result service
